Re-organized Model classes.
