+++ CodeChallenge/CodeChallenge.csproj
@@ -17,7 +17,6 @@
     <PackageReference Include="Refit" Version="5.0.23" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Models\" />
     <Folder Include="Common\" />
     <Folder Include="Effects\" />
   </ItemGroup>
