+++ CodeChallenge/Models/Response/GenreResponse.cs
@@ -1,0 +1,12 @@
+﻿using CodeChallenge.Models.Response.Data;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CodeChallenge.Models.Response
+{
+    public class GenreResponse
+    {
+        public List<Genre> Genres { get; set; }
+    }
+}
